chore: refactor codes for sync_cac_content
